+++ next-docs/pages/components/typography/caption.tsx
@@ -1,3 +1,4 @@
+import Link from 'next/link';
 import Preview from '../../../components/codePreview/Preview';
 import ComponentPageDescription from '../../../components/ComponentPageDescription';
 import PropsTable from '../../../components/PropsTable';
@@ -13,6 +14,15 @@ const PageCaption = () => {
   return (
     <>
       <ComponentPageDescription title="Caption">
+        <p className="font-medium">
+          NB! The component is depricated. Please use our{' '}
+          <Link href="/core/typography">
+            <a className="text-piccolo transition-colors duration-200 hover:text-hit visited:text-hit">
+              new Typography
+            </a>
+          </Link>{' '}
+          built with Tailwind.
+        </p>
         <p>
           By default Caption renders as Span tag with Bulma color and size of
           10px.