+++ packages/select/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @heathmont/moon-select
 
+## 9.8.1
+
+### Patch Changes
+
+- Fix code view
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.8.1
+  - @heathmont/moon-themes@9.8.1
+  - @heathmont/moon-utils@9.8.1
+
+## 9.8.0
+
+### Minor Changes
+
+- Deploy fixes
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.8.0
+  - @heathmont/moon-themes@9.8.0
+  - @heathmont/moon-utils@9.8.0
+
 ## 9.7.8
 
 ### Patch Changes